share fixture via conftest
