+++ cstar/tests/unit_tests/orchestration/conftest.py
@@ -8,7 +8,7 @@ from pathlib import Path
 import pytest
 
 from cstar.base.utils import additional_files_dir
-from cstar.orchestration.models import RomsMarblBlueprint, Step, Workplan
+from cstar.orchestration.models import Application, RomsMarblBlueprint, Step, Workplan
 
 
 @pytest.fixture
@@ -422,3 +422,45 @@ def default_blueprint_path() -> str:
     str
     """
     return "~/code/cstar/cstar/additional_files/templates/blueprint.yaml"
+
+
+@pytest.fixture
+def single_step_workplan(
+    tmp_path: Path,
+    bp_templates_dir: Path,
+) -> Workplan:
+    """Generate a workplan that contains a single step.
+
+    Parameters
+    ----------
+    tmp_path : Path
+        Temporary directory for test outputs
+    bp_templates_dir : Path
+        Fixture returning the path to the directory containing blueprint template files
+
+    Returns
+    -------
+    Workplan
+    """
+    bp_tpl_path = bp_templates_dir / "blueprint.yaml"
+    default_output_dir = "output_dir: ."
+
+    bp_path = tmp_path / "blueprint.yaml"
+    bp_content = bp_tpl_path.read_text()
+    bp_content = bp_content.replace(
+        default_output_dir, f"output_dir: {tmp_path.as_posix()}"
+    )
+    bp_content.replace(Application.SLEEP.value, Application.ROMS_MARBL.value)
+    bp_path.write_text(bp_content)
+
+    return Workplan(
+        name="single-step-workplan",
+        description="A workplan with a single step.",
+        steps=[
+            Step(
+                name="s-00",
+                application=Application.SLEEP.value,
+                blueprint=bp_path.as_posix(),
+            ),
+        ],
+    )
